split out commands.
Now each command gets their own little class.
It might be worth looking to see if any should be merged but at this stage just as is is fine.
Moved some shared elements into a shared folder.
