+++ src/YarnSpinner.Console/Shared/BasicBlockAnalysis.cs
