+++ src/YarnSpinner.Console/Shared/Log.cs
